+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.13]
+### Added
+- Match Note types as an enum
+- Parse partnership data in Match
+
 ## [0.0.12]
 ### Added
 - Parse Cricinfo season page to print all series in a season